requirements.txt
numpy==1.25.0
opencv_python==4.8.0.74
Pillow==9.5.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
